backend/requirements.txt
flask==3.0.3
flask_cors==4.0.1
pillow==10.4.0
tensorflow==2.17.0
